Create Pointing to your first note
Pointing to your first note
You're building a simple note-taking app. Each note will be stored in a file, so the first step is to create a File object that points to where your note will be saved. In this task, you'll prepare to manage a file called "note.txt".

Instructions

Import the File class from java.io.
Create a File object named noteFile that points to "note.txt".