Create Updating names in a contact list
Updating names in a contact list
A contact management system stores names in a list. When a name needs to be updated (e.g., correcting typos), you can use ListIterator to modify entries while iterating.

Instructions

Create a ListIterator object for the contacts list.
In the condition of the while loop, check if more elements are available in contacts list.
Retrieve next element.
If a name is Jon, set it to John instead.