Create Formatting and parsing dates
Formatting and parsing dates
Formatting and parsing dates are essential for working with different date representations. In this exercise, you will format a date and parse a string into a LocalDate.

Instructions

Retrieve the current date.
Define a DateTimeFormatter with the pattern dd-MM-yyyy.
Format the current date using the formatter.
Parse a date string using the defined formatter.